+++ src/test/java/randoop/main/gentests/classes/AnotherDependency.java
@@ -1,0 +1,4 @@
+package randoop.main.gentests.classes;
+
+/** Used in tests in AutoAddDependenciesTest */
+public class AnotherDependency {}