+++ src/test/java/randoop/main/gentests/classes/AnotherTestedClass.java
@@ -1,0 +1,7 @@
+package randoop.main.gentests.classes;
+
+/** Used in tests in AutoAddDependenciesTest */
+public class AnotherTestedClass {
+
+  public void method(AnotherDependency anotherDependency) {}
+}