+++ src/test/java/randoop/main/gentests/classes/TestedClass.java
@@ -1,0 +1,9 @@
+package randoop.main.gentests.classes;
+
+/** Used in tests in AutoAddDependenciesTest */
+public class TestedClass {
+
+  public TestedClass(ConstructorDependency constructorDependency) {}
+
+  public void method(MethodDependency methodDependency) {}
+}